django/requirements.txt
Django>=5.1
djangorestframework
django-filter
djangorestframework-simplejwt
firebase-admin
psycopg2-binary
pgvector
django-environ
google-generativeai
celery
redis
